hw7-refactor-packages: перераспределила все по папкам по примеру как в чате обсуждали
