+++ app/src/main/java/com/example/weather/domain/repo/weather/WeathersRepo.kt
@@ -1,4 +1,7 @@
-package com.example.weather.model
+package com.example.weather.domain.repo.weather
+
+import com.example.weather.domain.model.Weather
+import com.example.weather.domain.model.WeatherDTO
 
 interface WeathersRepo {
     fun getWeatherOfRusCities(): List<Weather>
